authentication tests almost implemented and query parameters added to getusers

+++ src/controllers/meal-controller.js
@@ -7,11 +7,11 @@ const controller = {
       if (err) {
         console.log(err);
         res.status(400).json({
-          Status: 400,
-          Message: `Something went wrong`,
+          status: 400,
+          message: `Something went wrong`,
         });
       } else {
-        res.send(rows);
+        res.status(200).json(rows);
       }
     });
   },
@@ -25,14 +25,14 @@ const controller = {
         if (err) {
           console.log(err);
           res.status(400).json({
-            Status: 400,
-            Message: `Something went wrong`,
+            status: 400,
+            message: `Something went wrong`,
           });
         } else {
           if (rows.length == 0) {
             res.status(404).json({
               Status: 404,
-              Message: `Meal not found`,
+              message: `Meal not found`,
             });
           } else {
             res.send(rows);
@@ -40,6 +40,64 @@ const controller = {
         }
       }
     );
+  },
+
+  validateMeal: (req, res, next) => {
+    let meal = req.body;
+
+    let {
+      name,
+      description,
+      isActive,
+      isVega,
+      isVegan,
+      isToTakeHome,
+      imageUrl,
+      allergenes,
+      maxAmountOfParticipants,
+      price,
+    } = meal;
+
+    try {
+      assert(name != null, "name must be provided in request");
+      assert(description != null, "description must be provided in request");
+
+      assert(isActive != null, "isActive must be provided in request");
+      assert(isVega != null, "isVega must be provided in request");
+
+      assert(isVegan != null, "isVegan must be provided in request");
+      assert(isToTakeHome != null, "isToTakeHome must be provided in request");
+
+      assert(imageUrl != null, "imageUrl must be provided in request");
+
+      assert(
+        maxAmountOfParticipants != null,
+        "maxAmountOfParticipants must be provided in request"
+      );
+      assert(price != null, "price must be provided in request");
+
+      assert(typeof name == "string", "name must be a string");
+      assert(typeof description == "string", "description must be a string");
+      assert(typeof isActive == "boolean", "isActive must be a boolean");
+      assert(typeof isVega == "boolean", "isVega must be a boolean");
+      assert(typeof isVegan == "boolean", "isVegan must be a boolean");
+      assert(
+        typeof isToTakeHome == "boolean",
+        "isToTakeHome must be a boolean"
+      );
+      assert(typeof imageUrl == "string", "imageUrl must be a string");
+      assert(
+        typeof maxAmountOfParticipants == "number",
+        "maxAmountOfParticipants must be a number"
+      );
+      assert(typeof price == "number", "price must be a number");
+      next();
+    } catch (err) {
+      res.status(400).json({
+        status: 400,
+        message: err.message,
+      });
+    }
   },
 
   addMeal: (req, res) => {
@@ -52,7 +110,7 @@ const controller = {
     } else {
       const date = new Date().toISOString();
       Database.query(
-        "INSERT INTO `meal`(`isActive`, `isVega`, `isVegan`, `isToTakeHome`, `dateTime`, `maxAmountOfParticipants`, `price`, `imageUrl`, `cookId`, `createDate`, `updateDate`, `name`, `description`, `allergenes`) VALUES (?,?,?,?,?,?,?,?,?,?,?,?,?,?); SELECT  * from `meal` WHERE id = LAST_INSERT_ID();",
+        "INSERT INTO `meal`(`isActive`, `isVega`, `isVegan`, `isToTakeHome`, `dateTime`, `maxAmountOfParticipants`, `price`, `imageUrl`, `cookId`, `createDate`, `updateDate`, `name`, `description`) VALUES (?,?,?,?,?,?,?,?,?,?,?,?,?); SELECT  * from `meal` WHERE id = LAST_INSERT_ID();",
         [
           meal.isActive,
           meal.isVegan,
@@ -67,14 +125,14 @@ const controller = {
           date,
           meal.name,
           meal.description,
-          meal.allergenes,
+          // meal.allergenes,
         ],
         (err, rows, fields) => {
           if (err) {
             console.log(err);
             res.status(400).json({
-              Status: 400,
-              Message: `Something went wrong`,
+              status: 400,
+              message: `Something went wrong`,
             });
           } else {
             res.status(201).json({
@@ -95,8 +153,8 @@ const controller = {
         if (err) {
           console.log(err);
           res.status(400).json({
-            Status: 400,
-            Message: `Something went wrong`,
+            status: 400,
+            message: `Something went wrong`,
           });
         } else {
           res.send(rows);
@@ -107,24 +165,58 @@ const controller = {
 
   deleteMealById: (req, res) => {
     const mealId = req.params.mealId;
-    Database.query(
-      "DELETE FROM meal WHERE id = ?",
-      [mealId],
-      (err, rows, fields) => {
-        if (err) {
-          console.log(err);
-          res.status(400).json({
-            Status: 400,
-            Message: `Something went wrong`,
-          });
-        } else {
-          res.status(200).json({
-            Status: 200,
-            Message: `succesfully deleted meal`,
-          });
+    if (typeof req.jwtUserId === "undefined") {
+      res.status(400).json({
+        Status: 400,
+        message: `Not logged in`,
+      });
+    } else {
+      Database.query(
+        "SELECT cookId FROM meal WHERE id = ?; SELECT cookId FROM meal WHERE id = ? AND cookId = ?",
+        [mealId, mealId, req.jwtUserId],
+        (err, rows, fields) => {
+
+          if (err) {
+            console.log(err);
+            res.status(400).json({
+              status: 400,
+              message: `Something went wrong`,
+            });
+          } else {
+            if (rows[0].length == 0) {
+              return res.status(404).json({
+                statusCode: 404,
+                message: `Meal does not exist`,
+              });
+            } else if (rows[1].length == 0) {
+              return res.status(403).json({
+                status: 403,
+                message: `Unauthorized`,
+              });
+            } else {
+              Database.query(
+                "DELETE FROM meal WHERE id = ?",
+                [mealId],
+                (err, rows, fields) => {
+                  if (err) {
+                    console.log(err);
+                    res.status(400).json({
+                      status: 400,
+                      message: `Something went wrong`,
+                    });
+                  } else {
+                    res.status(200).json({
+                      status: 200,
+                      message: `succesfully deleted meal`,
+                    });
+                  }
+                }
+              );
+            }
+          }
         }
-      }
-    );
+      );
+    }
   },
 
   enrollMeal: (req, res) => {
@@ -136,13 +228,13 @@ const controller = {
         if (err) {
           console.log(err);
           res.status(400).json({
-            Status: 400,
-            Message: `Something went wrong`,
+            status: 400,
+            message: `Something went wrong`,
           });
         } else {
           res.status(200).json({
-            Status: 200,
-            Message: `succesfully enrolled into meal`,
+            status: 200,
+            message: `succesfully enrolled into meal`,
           });
         }
       }
